Create hover-intent mixin
